pyproject.toml
[tool.poetry]
name = "flask-db"
version = "0.1.0"
description = ""
authors = ["Nikita Demakov<demakoffnik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
